searcch_backend/api/ticket_creation/antAPI/requirements.txt
astroid==2.12.12
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
dill==0.3.5.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.5
python-dotenv==0.21.0
SQLAlchemy==1.4.42
tomli==2.0.1
tomlkit==0.11.5
Werkzeug==2.2.2
wrapt==1.14.1
